+++ app/assets/javascripts/views/story_view.js
@@ -494,8 +494,12 @@ module.exports = FormView.extend({
         );
     }
 
-    const $storyEstimate = this.$('[data-story-estimate]');
-    if ($storyEstimate.length) {
+    this.renderSelects();
+  },
+
+  renderSelects: function() {
+    const $storyEstimateSelect = this.$('[data-story-estimate]');
+    if ($storyEstimateSelect.length) {
       const storyEstimateOptions = this.model.point_values().map(this.createStoryEstimateOptions);
       ReactDOM.render(
         <StorySelect
@@ -506,10 +510,10 @@ module.exports = FormView.extend({
           selected={this.model.get('estimate')}
           disabled={this.model.notEstimable() || this.isReadonly()}
         />,
-        $storyEstimate.get(0)
+        $storyEstimateSelect.get(0)
       );
 
-      this.bindElementToAttribute($storyEstimate.find('select[name="estimate"]'), 'estimate');
+      this.bindElementToAttribute($storyEstimateSelect.find('select[name="estimate"]'), 'estimate');
     }
 
     const $storyTypeSelect = this.$('[data-story-type]');